requirements.txt
# ============================================================================
# SPTrans Pipeline - Python Dependencies
# ============================================================================
# Descrição: Todas as dependências necessárias para o pipeline
# Python Version: 3.9+
# ============================================================================

# ====================
# CORE
# ====================
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ====================
# DATA PROCESSING
# ====================
# PySpark
pyspark==3.5.0
delta-spark==3.0.0

# Pandas
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# ====================
# DATABASES
# ====================
# PostgreSQL
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Redis
redis==5.0.1

# ====================
# API & HTTP
# ====================
requests==2.31.0
urllib3==2.1.0

# ====================
# AIRFLOW
# ====================
apache-airflow==2.8.0
apache-airflow-providers-apache-spark==4.5.0
apache-airflow-providers-postgres==5.9.0
apache-airflow-providers-redis==3.5.0
apache-airflow-providers-http==4.7.0

# ====================
# AWS/S3/MinIO
# ====================
boto3==1.34.10
s3fs==2024.2.0

# ====================
# MONITORING
# ====================
# Prometheus
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7

# ====================
# GEOSPATIAL
# ====================
# Geocoding
geopy==2.4.1
Shapely==2.0.2

# ====================
# DATA QUALITY
# ====================
great-expectations==0.18.8

# ====================
# TESTING
# ====================
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-xdist==3.5.0

# ====================
# CODE QUALITY
# ====================
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3
isort==5.13.2

# ====================
# DOCUMENTATION
# ====================
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# ====================
# UTILITIES
# ====================
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
tabulate==0.9.0
tqdm==4.66.1

# ====================
# DEVELOPMENT
# ====================
ipython==8.19.0
jupyter==1.0.0
jupyterlab==4.0.10

# ====================
# OPTIONAL
# ====================
# Kafka (se usar streaming real)
# kafka-python==2.0.2

# Trino (se usar para queries distribuídas)
# trino==0.328.0

# ============================================================================
# INSTALLATION INSTRUCTIONS
# ============================================================================
# 
# Development:
#   pip install -r requirements.txt
#
# Production (sem dev tools):
#   pip install -r requirements.txt --no-dev
#
# Docker:
#   RUN pip install --no-cache-dir -r requirements.txt
#
# ============================================================================
